props back to monolit version

+++ src/main/java/com/exrates/inout/properties/models/EthereumTokenProperty.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 @Data
 @Builder(builderClassName = "Builder", toBuilder = true)
@@ -19,5 +19,5 @@ public class EthereumTokenProperty {
     private String contract;
     private boolean isERC20;
     private ExConvert.Unit unit;
-    private BigDecimal minWalletBalance;
+    private BigInteger minWalletBalance;
 }
